🚚 `NavigationView` => moved in dedicated file
