+++ src/lib/views/default-layout.view.ts
@@ -1,4 +1,5 @@
 import { AnyVirtualDOM, ChildrenLike, VirtualDOM } from '@youwol/rx-vdom'
+import { NavigationView } from './navigation.view'
 import { Router } from '../router'
 import { ImmutableTree } from '@youwol/rx-tree-views'
 import { NavigationHeader } from './navigation.view'
@@ -81,33 +82,6 @@ export class DefaultLayoutView implements VirtualDOM<'div'> {
                     },
                 ],
             },
-        ]
-    }
-}
-
-export class NavigationView implements VirtualDOM<'div'> {
-    public readonly router: Router
-
-    public readonly tag = 'div'
-    public readonly class = 'h-100 w-100 overflow-auto'
-    public readonly children: ChildrenLike
-    public readonly style = {
-        fontSize: '0.9rem',
-    }
-    constructor(params: { router: Router }) {
-        Object.assign(this, params)
-
-        this.children = [
-            new ImmutableTree.View({
-                state: this.router.explorerState,
-                headerView: (explorerState, node) => {
-                    return new NavigationHeader({
-                        node,
-                        router: this.router,
-                        explorerState,
-                    })
-                },
-            }),
         ]
     }
 }
